checkout works
