+++ controllers/api/cartRoutes.js
@@ -20,25 +20,5 @@ router.post('/cart/add', withAuth, async (req, res) => {
 
 
 
-router.post('/add', withAuth, async (req, res) => {
-    try {
-      console.log("Request body: ", req.body);  
-      const { bookId, sellerId } = req.body;
-      const buyerId = req.session.user_id; 
-      const buyerEmail = req.session.email;
-      const transaction = await Transaction.create({
-        book_id: bookId,
-        seller_id: sellerId,
-        buyer_id: buyerId,
-        buyer_email: buyerEmail,
-        status: 'active',
-        transaction_date: new Date(),
-      });
-      res.status(200).json({ message: 'Book added to cart and seller notified.' });
-    } catch (err) {
-      console.error(err);
-      res.status(500).json({ error: 'Failed to add book to cart.' });
-    }
-  });
 
 module.exports = router;
